Fix random memory corruption due to buggy list traversal
The implementation uses list_for_each_entry_safe(), which caches the
head->next pointer locally so the current node can be deleted. But
since the lock is released in the body of the loop, the object pointed
to by the cached ->next pointer may be changed or freed by another
thread before we dereference it.

Because kfree() can be called from any kernel context, it is actually
entirely unnecessary to release the lock in the loop body. This patch
fixes the bug by holding the lock for the whole list traversal.

This is only called during driver teardown, but should still be fixed.

Signed-off-by: Calvin Owens <calvin@quilt.com>