Set CONFIG_REASSOCIATION=n in out of tree IW612 driver
This comment is from woal_reassociation_thread():

    /* Reassociation should be disabled completely by application
       if wlan_set_user_scan_ioctl/wlan_set_wap is used. */

No upstream driver has this periodic reassociation logic: since the
userspace bits take care of that, we don't want it. Disable it.

Also, add a couple WARNINGs to log stacktraces at places in the driver
which I suspect might be implicated in other failures.

Also, drop two warnings that break the build with LLVM.

Signed-off-by: Calvin Owens <calvin@quilt.com>